test: add functional test for user register
- add test enviroment -> compose, seed n migrations
