+++ api/env.example
@@ -1,0 +1,3 @@
+DATABASE_URL_DEV=your_db_dev_url
+DATABASE_URL_TEST=your_db_test_url
+MODE=test
