+++ api/package.json
@@ -3,7 +3,9 @@
   "version": "1.0.0",
   "main": "index.js",
   "scripts": {
-    "dev": "ts-node-dev --respawn --transpile-only -r tsconfig-paths/register src/start.ts"
+    "dev": "ts-node-dev --respawn --transpile-only -r tsconfig-paths/register src/start.ts",
+    "migrate": "ts-node-dev src/libs/migration.ts",
+    "seed": "npx prisma db seed"
   },
   "prisma": {
     "seed": "ts-node prisma/seed.ts"
@@ -25,6 +27,7 @@
     "multer": "^2.0.2"
   },
   "devDependencies": {
+    "@faker-js/faker": "^10.1.0",
     "@types/bcrypt": "^6.0.0",
     "@types/cookie-parser": "^1.4.9",
     "@types/cors": "^2.8.19",
