--- api/src/utils/PrismaHelper.ts
@@ -1,3 +1,0 @@
-export function buildNestedCreate<T>(data?: T[]) {
-	return data && data.length > 0 ? { create: data } : undefined;
-}
