+++ api/tsconfig.json
@@ -40,6 +40,6 @@
       "@/*": [
         "./src/*"
       ]
-    }
+     }
   }
 }
