+++ app/cypress/e2e/auth/login.cy.ts
@@ -21,18 +21,18 @@ describe('Login test', () => {
 
 	// email
 	it('ci username empty', () => {
-  	cy.get('[name="password"]').clear();
-  	cy.get('[name="password"]').type(Cypress.env('PASSWORD'));
-  	cy.get('#root button.w-full').click();
-    cy.get('#root p.text-red-500').should('have.text', 'El correo electrónico no es válido');
+	cy.get('[name="password"]').clear();
+	cy.get('[name="password"]').type(Cypress.env('PASSWORD'));
+	cy.get('#root button.w-full').click();
+	cy.get('#root p.text-red-500').should('have.text', 'El correo electrónico no es válido');
 	});
 
 	it('ci username not exits', () => {
-  	cy.get('[name="email"]').clear();
-  	cy.get('[name="email"]').type('nouser@eexample.com');
-  	cy.get('[name="password"]').clear();
-  	cy.get('[name="password"]').type(Cypress.env('PASSWORD'));
-  	cy.get('#root button.w-full').click();
+	cy.get('[name="email"]').clear();
+	cy.get('[name="email"]').type('nouser@eexample.com');
+	cy.get('[name="password"]').clear();
+	cy.get('[name="password"]').type(Cypress.env('PASSWORD'));
+	cy.get('#root button.w-full').click();
    cy.get('#root div.Toastify__toast').should('have.text', 'Usuario o contraseña no válidos.');
 
 	});
@@ -42,16 +42,16 @@ describe('Login test', () => {
 	  cy.get('[name="email"]').clear();
 		cy.get('[name="email"]').type(Cypress.env('ADMIN'));
 		cy.get('[name="password"]').clear();
- 	  cy.get('#root button.w-full').click();
+	  cy.get('#root button.w-full').click();
 
-    cy.get('#root p.text-red-500').should('have.text', 'Ingresa la contraseña.');
+	cy.get('#root p.text-red-500').should('have.text', 'Ingresa la contraseña.');
 	});
 
 	it('password-ci-bad-password', () => {
-  	cy.get('[name="email"]').clear();
-  	cy.get('[name="email"]').type(Cypress.env('ADMIN'));
-  	cy.get('[name="password"]').clear();
-  	cy.get('[name="password"]').type('123456789');
+	cy.get('[name="email"]').clear();
+	cy.get('[name="email"]').type(Cypress.env('ADMIN'));
+	cy.get('[name="password"]').clear();
+	cy.get('[name="password"]').type('123456789');
 		cy.get('#root button.w-full').click();
 		cy.get('#root div.Toastify__toast').should('have.text', 'Usuario o contraseña no válidos.');
 	});
